arm_gicv2m: Add GICv2m widget to support MSIs
The ARM GICv2m widget is a little device that handle MSI interrupt
writes to a trigger register and ties them to a range of interrupt lines
wires to the GIC.  It has a few status/id registers and the interrupt wires,
and that's about it.

A board instantiates the device by setting the base SPI number and
number SPIs for the frame.  The base-spi parameter is indexed in the SPI
number space only, so base-spi == 0, means IRQ number 32.  Writes to the
trigger register write the GIC IRQ number, so we have to subtract 32
from that and then index into our frame of SPIs.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>