kvm_stat: Fix filter specifier
Check the return value from re.match against its boolean value and not
whether it returned a match object or not (this may have changed in
recent python versions).

Also honor the filter request in outputs making it much easier to use
this tool to gather crucial statistics for a given workload.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>